Index page and feed page adjustments
Fixed the index page question display to be reactive.

Adjusted the resources sections to be thinner. Spread the open full button across the entire resources box.

Linked the user activity feed to the individual posts they are displaying. User is now able to click and be redirected to the appropriate post page.